Add support to keep specific nodes
We might want to keep some nodes in the final result even if the area is
below the threshold.

In the CSV file, we can specify if the point should be excluded from the
simplification. The --keep-col option specify this column. A value of 0
indicates the node could be simplified and 1 indicates that a node is
preserved in the final output.

Signed-off-by: Francis Giraldeau <francis.giraldeau@nrc-cnrc.gc.ca>